Move log to pkg/log.
