+++ cmd/registry/cmd/compute/lintstats/lintstats.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/apigee/registry/cmd/registry/types"
 	"github.com/apigee/registry/gapic"
-	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
+	"github.com/apigee/registry/pkg/log"
 	"github.com/apigee/registry/pkg/names"
 	"github.com/apigee/registry/pkg/visitor"
 	"github.com/apigee/registry/rpc"
