+++ cmd/registry/cmd/vocabulary/difference.go
@@ -17,8 +17,8 @@ package vocabulary
 import (
 	"fmt"
 
-	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
+	"github.com/apigee/registry/pkg/log"
 	"github.com/google/gnostic/metrics/vocabulary"
 	"github.com/spf13/cobra"
 	"google.golang.org/protobuf/encoding/protojson"
