+++ cmd/registry/controller/controller.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/apigee/registry/cmd/registry/patterns"
-	"github.com/apigee/registry/log"
+	"github.com/apigee/registry/pkg/log"
 	"github.com/apigee/registry/rpc"
 )
