+++ cmd/registry/main.go
@@ -20,7 +20,7 @@ import (
 	"os"
 
 	"github.com/apigee/registry/cmd/registry/cmd"
-	"github.com/apigee/registry/log"
+	"github.com/apigee/registry/pkg/log"
 	"github.com/google/uuid"
 )
