+++ cmd/registry/patch/apply.go
@@ -25,8 +25,8 @@ import (
 	"strings"
 
 	"github.com/apigee/registry/cmd/registry/tasks"
-	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
+	"github.com/apigee/registry/pkg/log"
 	"gopkg.in/yaml.v3"
 )
