--- log/README.md
@@ -1,5 +1,0 @@
-# log
-
-This package provides logging for the Registry project. See
-https://pkg.go.dev/github.com/apigee/registry/log for package documentation and
-examples.
