+++ pkg/log/README.md
@@ -1,0 +1,5 @@
+# log
+
+This package provides logging for the Registry project. See
+https://pkg.go.dev/github.com/apigee/registry/pkg/log for package documentation and
+examples.
