+++ pkg/log/interceptor/interceptor.go
@@ -21,7 +21,7 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/apigee/registry/log"
+	"github.com/apigee/registry/pkg/log"
 	"github.com/google/uuid"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
