+++ pkg/wipeout/tasks.go
@@ -17,8 +17,8 @@ package wipeout
 import (
 	"context"
 
-	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
+	"github.com/apigee/registry/pkg/log"
 	"github.com/apigee/registry/rpc"
 )
